requirements.txt
matplotlib==3.4.3
numpy==1.20.3
pandas==1.3.3
Pillow==8.4.0
progress==1.6
pyreadr==0.4.2
scikit_learn==1.0.2
scipy==1.7.1
seaborn==0.11.2
torch==1.8.1
torchvision==0.2.0
